+++ src/Service/Shopware/DataExtractor/OrderAddressDataExtractor.php
@@ -2,6 +2,8 @@
 
 namespace PostNL\Shopware6\Service\Shopware\DataExtractor;
 
+use PostNL\Shopware6\Defaults;
+use Shopware\Core\Checkout\Order\Aggregate\OrderAddress\OrderAddressCollection;
 use Shopware\Core\Checkout\Order\Aggregate\OrderAddress\OrderAddressEntity;
 use Shopware\Core\System\Country\CountryEntity;
 
@@ -17,5 +19,13 @@ class OrderAddressDataExtractor
 
         // TODO Exception
         throw new \Exception('Could not extract country');
+    }
+
+    public function filterByAddressType(OrderAddressCollection $addresses, string $addressType = "01"): OrderAddressCollection
+    {
+        /** @var OrderAddressCollection */
+        return $addresses->filter(function(OrderAddressEntity $address) use ($addressType) {
+            return $address->getCustomFields()[Defaults::CUSTOM_FIELDS_KEY]['addressType'] === $addressType;
+        });
     }
 }