+++ src/Subscriber/AddressSubscriber.php
@@ -2,6 +2,7 @@
 
 namespace PostNL\Shopware6\Subscriber;
 
+use PostNL\Shopware6\Defaults;
 use Shopware\Core\Checkout\Customer\CustomerEvents;
 use Shopware\Core\Framework\Event\DataMappingEvent;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
@@ -12,9 +13,9 @@ class AddressSubscriber implements EventSubscriberInterface
     public static function getSubscribedEvents(): array
     {
         return [
-            CustomerEvents::MAPPING_REGISTER_ADDRESS_BILLING => 'onMappingAddress',
+            CustomerEvents::MAPPING_REGISTER_ADDRESS_BILLING  => 'onMappingAddress',
             CustomerEvents::MAPPING_REGISTER_ADDRESS_SHIPPING => 'onMappingAddress',
-            CustomerEvents::MAPPING_ADDRESS_CREATE => 'onMappingAddress',
+            CustomerEvents::MAPPING_ADDRESS_CREATE            => 'onMappingAddress',
         ];
     }
 
@@ -22,15 +23,14 @@ class AddressSubscriber implements EventSubscriberInterface
     {
         $inputData = $event->getInput();
         $outputData = $event->getOutput();
-//        dd($inputData);
-        $postNL = [
-            'streetName'=>$inputData->get('streetName'),
-            'houseNumber'=>$inputData->get('houseNumber'),
-            'houseNumberAddition'=>$inputData->get('houseNumberAddition'),
-        ];
-        $outputData['customFields'] = [
-            'postNL' => $postNL,
-        ];
+
+        $outputData['customFields'] = array_merge_recursive($outputData['customFields'] ?? [], [
+            Defaults::CUSTOM_FIELDS_KEY => [
+                'streetName'          => $inputData->get('streetName'),
+                'houseNumber'         => $inputData->get('houseNumber'),
+                'houseNumberAddition' => $inputData->get('houseNumberAddition'),
+            ],
+        ]);
 
         $event->setOutput($outputData);